chore(ui-react): lint components/Geo directory
