+++ packages/react/.eslintrc.js
@@ -70,6 +70,7 @@ module.exports = {
         // TODO Uncomment the below lines on an individual basis as the updates to pass linting are completed.
         // If needed, these can be broken down further, e.g. 'src/components/Geo/**/*'
         'src/helpers/**/*',
+        'src/components/Geo/**/*',
         // 'src/components/**/*',
         // 'src/hooks/**/*',
         // 'src/primitives/**/*',
